Typo fix & Linting

+++ lua/modules/crossfire/src/compile.lua
@@ -15,39 +15,31 @@ function compile.file_exists(name)
 end
 
 function compile.baseName()
-	local baseName
-	baseName = config.moduleDir:gsub("/scripts/","")
-	baseName = baseName:gsub("/","")
-	return baseName
+    local baseName
+    baseName = config.moduleDir:gsub("/scripts/", "")
+    baseName = baseName:gsub("/", "")
+    return baseName
 end
 
 function compile.loadScript(script)
 
+    local cachefile
+    cachefile = moduleDir .. "compiled/" .. script:gsub("/", "_") .. "c"
 
-	local cachefile
-	cachefile = moduleDir .. "compiled/" .. script:gsub("/", "_") .. "c"
+    if compile.file_exists("/scripts/" .. compile.baseName() .. ".nocompile") == true then config.useCompiler = false end
 
-
-    if compile.file_exists("/scripts/" .. compile.baseName() .. ".nocompile" ) == true then
-		config.useCompiler = false
-	end
-
-    if compile.file_exists("/scripts/nocompile" ) == true then
-		config.useCompiler = false
-	end
+    if compile.file_exists("/scripts/nocompile") == true then config.useCompiler = false end
 
     if config.useCompiler == true then
         if compile.file_exists(cachefile) ~= true then
             system.compile(script)
             os.rename(script .. 'c', cachefile)
         end
-        --print("Loading: " .. cachefile)
+        -- print("Loading: " .. cachefile)
         return loadfile(cachefile)
     else
-        if compile.file_exists(cachefile) == true then
-            os.remove(cachefile)
-        end		
-		--print("Loading: " .. script)
+        if compile.file_exists(cachefile) == true then os.remove(cachefile) end
+        -- print("Loading: " .. script)
         return loadfile(script)
     end
